+++ dsmr_notification/views.py
@@ -1,1 +1,0 @@
-# Just here to support the dsmr_notification app