+++ dsmr_stats/tests/test_management_command.py
@@ -8,7 +8,9 @@ from dsmr_backend.tests.mixins import InterceptStdoutMixin
 
 class TestManagementCommand(InterceptStdoutMixin, TestCase):
     def test_dsmr_stats_clear_statistics(self):
-        with self.assertRaisesMessage(CommandError, 'Intended usage is NOT production! Force by using --ack-to-delete-my-data'):
+        expected_error = 'Intended usage is NOT production! Force by using --ack-to-delete-my-data'
+
+        with self.assertRaisesMessage(CommandError, expected_error):
             self._intercept_command_stdout('dsmr_stats_clear_statistics')
 
         with mock.patch('dsmr_stats.services.clear_statistics') as service_mock: