+++ test-all.sh
@@ -1,15 +1,19 @@
 #!/bin/bash
 
 echo ""
+echo "--- Running Pylama for code audit..."
+pylama
+
+echo ""
 echo "--- Testing with SQLite..."
-pytest --cov --cov-report=html:coverage_report/html --cov-report=term --ds=dsmrreader.config.test_sqlite 
+pytest --pylama --cov --cov-report=html:coverage_report/html --cov-report=term --ds=dsmrreader.config.test_sqlite 
 
 
 echo ""
 echo "--- Testing with PostgreSQL..."
-pytest --cov --cov-report=html:coverage_report/html --cov-report=term --ds=dsmrreader.config.test_postgresql
+pytest --pylama --cov --cov-report=html:coverage_report/html --cov-report=term --ds=dsmrreader.config.test_postgresql
 
 
 echo ""
 echo "--- Testing with MySQL..."
-pytest --cov --cov-report=html:coverage_report/html --cov-report=term --ds=dsmrreader.config.test_mysql
+pytest --pylama --cov --cov-report=html:coverage_report/html --cov-report=term --ds=dsmrreader.config.test_mysql